refactor(utils): add utils.desktop
The AppServerThread and ClientWindowProcess were both moved into this file.
